+++ src/commands/subscription/raid/collect_option.js
@@ -10,11 +10,13 @@ module.exports = (WDR, Functions, source, oMessage, bMessage, Member, gym_name_a
 
     collector.on('collect', CollectedMsg => {
 
-        try {
-            CollectedMsg.delete();
-        // eslint-disable-next-line no-empty
-        } catch (e) {
+        if (!CollectedMsg.content.startsWith(WDR.Config.PREFIX)) {
+            try {
+                CollectedMsg.delete();
+            // eslint-disable-next-line no-empty
+            } catch (e) {
 
+            }
         }
 
         let input = CollectedMsg.content.split(' ')[0].toString().toLowerCase();
@@ -49,6 +51,8 @@ module.exports = (WDR, Functions, source, oMessage, bMessage, Member, gym_name_a
             case resume_words.some(word => input.includes(word)):
                 collector.stop('resume');
                 break;
+            default:
+                collector.stop('cancel');
         }
     });
